ENH: Move IndexingError to error/__init__.py per GH27656
